@@ -41,29 +42,26 @@
   },
   "homepage": "https://github.com/blackflux/node-tdd#readme",
   "devDependencies": {
-    "@babel/cli": "7.14.8",
-    "@babel/core": "7.15.0",
-    "@babel/register": "7.15.3",
-    "@blackflux/eslint-plugin-rules": "2.0.3",
-    "@blackflux/robo-config-plugin": "5.3.0",
-    "aws-sdk": "2.981.0",
-    "aws-sdk-wrap": "9.6.0",
-    "axios": "0.24.0",
-    "babel-eslint": "10.1.0",
-    "babel-preset-latest-node": "5.5.1",
-    "chai": "4.3.4",
+    "@babel/core": "7.17.9",
+    "@babel/eslint-parser": "7.17.0",
+    "@babel/register": "7.17.7",
+    "@blackflux/eslint-plugin-rules": "2.1.0",
+    "@blackflux/robo-config-plugin": "7.7.6",
+    "aws-sdk": "2.1112.0",
+    "aws-sdk-wrap": "12.0.4",
+    "axios": "0.26.1",
+    "c8": "7.11.0",
+    "chai": "4.3.6",
     "coveralls": "3.1.1",
-    "eslint": "7.32.0",
-    "eslint-config-airbnb-base": "14.2.1",
-    "eslint-plugin-import": "2.24.2",
+    "eslint": "8.13.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-json": "3.1.0",
-    "eslint-plugin-markdown": "2.2.0",
-    "eslint-plugin-mocha": "9.0.0",
-    "fancy-log": "1.3.3",
-    "js-gardener": "3.0.3",
-    "lambda-monitor-logger": "3.0.0",
-    "nyc": "15.1.0",
-    "semantic-release": "17.4.7"
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-mocha": "10.0.3",
+    "fancy-log": "2.0.0",
+    "js-gardener": "3.0.5",
+    "lambda-monitor-logger": "3.2.2"
   },
   "licenses": [
     {
@@ -72,51 +70,21 @@
     }
   ],
   "engines": {
-    "node": ">= 12"
-  },
-  "nyc": {
-    "exclude": [
-      "gardener.js",
-      "node_modules/*",
-      "coverage/*",
-      "lib/*"
-    ],
-    "tempDir": "./coverage/.nyc_output",
-    "report-dir": "./coverage",
-    "check-coverage": true,
-    "per-file": false,
-    "lines": 100,
-    "statements": 100,
-    "functions": 100,
-    "branches": 100,
-    "include": [
-      "**/*.js"
-    ],
-    "reporter": [
-      "lcov",
-      "text-summary"
-    ],
-    "require": [
-      "@babel/register"
-    ],
-    "extension": [],
-    "cache": true,
-    "all": true,
-    "babel": true
+    "node": ">= 14"
   },
   "files": [
     "lib"
   ],
   "dependencies": {
-    "callsites": "3.1.0",
+    "callsites": "4.0.0",
     "compare-urls": "2.0.0",
-    "joi-strict": "2.0.0",
+    "joi-strict": "2.0.1",
     "lodash.clonedeep": "4.5.0",
     "lodash.get": "4.4.2",
-    "minimist": "1.2.5",
-    "nock": "13.1.3",
-    "object-scan": "17.0.0",
-    "smart-fs": "2.0.2",
+    "minimist": "1.2.6",
+    "nock": "13.2.4",
+    "object-scan": "18.0.1",
+    "smart-fs": "3.0.1",
     "timekeeper": "2.2.0",
     "tmp": "0.2.1",
     "uuid": "8.3.0",
