@@ -1,6 +1,10 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-const gardener = require('js-gardener');
+/* eslint-disable import/no-extraneous-dependencies */
+import gardener from 'js-gardener';
+import fs from 'smart-fs';
+import process from 'process';
 
-if (require.main === module) {
-  gardener().catch(() => process.exit(1));
+if (process.argv[1] === fs.filename(import.meta.url)) {
+  gardener({
+    skip: []
+  }).catch(() => process.exit(1));
 }
