@@ -1,6 +1,13 @@
-module.exports.describe = require('./util/desc');
-module.exports.RequestRecorder = require('./modules/request-recorder');
-module.exports.EnvManager = require('./modules/env-manager');
-module.exports.TimeKeeper = require('./modules/time-keeper');
-module.exports.LogRecorder = require('./modules/log-recorder');
-module.exports.RandomSeeder = require('./modules/random-seeder');
+import describe_ from './util/desc.js';
+import RequestRecorder_ from './modules/request-recorder.js';
+import EnvManager_ from './modules/env-manager.js';
+import TimeKeeper_ from './modules/time-keeper.js';
+import LogRecorder_ from './modules/log-recorder.js';
+import RandomSeeder_ from './modules/random-seeder.js';
+
+export const describe = describe_;
+export const RequestRecorder = RequestRecorder_;
+export const EnvManager = EnvManager_;
+export const TimeKeeper = TimeKeeper_;
+export const LogRecorder = LogRecorder_;
+export const RandomSeeder = RandomSeeder_;
