@@ -1,11 +1,11 @@
-const assert = require('assert');
-const crypto = require('crypto');
-const get = require('lodash.get');
-const Joi = require('joi-strict');
+import assert from 'assert';
+import crypto from 'crypto';
+import get from 'lodash.get';
+import Joi from 'joi-strict';
 
 const libs = { Math, crypto };
 
-module.exports = (opts) => {
+export default (opts) => {
   Joi.assert(opts, Joi.object().keys({
     seed: Joi.string(),
     reseed: Joi.boolean()
