@@ -1,15 +1,15 @@
-const expect = require('chai').expect;
-const index = require('../src/index');
+import { expect } from 'chai';
+import * as index from '../src/index.js';
 
 describe('Testing index.js', () => {
   it('Testing exports', () => {
     expect(Object.keys(index)).to.deep.equal([
-      'describe',
-      'RequestRecorder',
       'EnvManager',
-      'TimeKeeper',
       'LogRecorder',
-      'RandomSeeder'
+      'RandomSeeder',
+      'RequestRecorder',
+      'TimeKeeper',
+      'describe'
     ]);
   });
 });
