@@ -1,8 +1,8 @@
-const expect = require('chai').expect;
-const util = require('../../../src/modules/request-recorder/util');
+import { expect } from 'chai';
+import { nullAsString } from '../../../src/modules/request-recorder/util.js';
 
 describe('Testing request-recorder/util.js', () => {
   it('Testing null is String', async () => {
-    expect(util.nullAsString(null)).to.equal('null');
+    expect(nullAsString(null)).to.equal('null');
   });
 });
